+++ public/gen-filename-json.ps1
@@ -1,4 +1,4 @@
-$files = Get-ChildItem "C:\Users\Naglis\Desktop\GITHUBO_REPOS\vbe-sort\public\2021" -Filter *.png
+$files = Get-ChildItem ".\" -Filter *.png
 $json = $files | ForEach-Object { 
    [ordered]@{ 
         filename = $_.Name