+++ src/mainPage/MathProblemsSection.tsx
@@ -2,14 +2,14 @@ import { Accordion, Form } from "react-bootstrap";
 import TopicSection from "./TopicSection";
 import topics from "./topics-names-list.json";
 import { useState } from "react";
+import allYearList from "./year-list.json";
+
+import ShuffleBar from "./ShuffleBar";
 
 export default function MathProblemsSection() {
-  const [yearList, setYearList] = useState<string[]>([
-    "2017g",
-    "2021g",
-    "2022g",
-    "2022k",
-  ]);
+  const [yearList, setYearList] = useState<string[]>(
+    allYearList.filter((year) => year !== "2023k" && year !== "2023g")
+  );
 
   const toggleYearInList = (yearToToggle: string) => {
     if (yearList.includes(yearToToggle)) {
@@ -18,6 +18,11 @@ export default function MathProblemsSection() {
       setYearList([...yearList, yearToToggle]);
     }
   };
+
+  //Shuffle stuff...
+  const [isShuffleOn, setShuffleOn] = useState<boolean>(true);
+  const reshuffle = () => {};
+  const undoShuffle = () => {};
 
   return (
     <div>
@@ -30,52 +35,37 @@ export default function MathProblemsSection() {
           sesijos), bei vieną mokykloje išspręsti kaip bandomąjį (greičiausiai
           2023 m. pakartotinės sesijos)
         </p>
-        <div style={{ marginTop: "20px" }}>
-          <Form>
-            <Form.Check
-              inline
-              label="2017 pag."
-              checked={yearList.includes("2017g")}
-              onChange={() => toggleYearInList("2017g")}
-            />
-            <Form.Check
-              inline
-              label="2021 pag."
-              checked={yearList.includes("2021g")}
-              onChange={() => toggleYearInList("2021g")}
-            />
-            <Form.Check
-              inline
-              label="2022 pag."
-              checked={yearList.includes("2022g")}
-              onChange={() => toggleYearInList("2022g")}
-            />
-            <Form.Check
-              inline
-              label="2022 pak."
-              checked={yearList.includes("2022k")}
-              onChange={() => toggleYearInList("2022k")}
-            />
-            <Form.Check
-              inline
-              label="2023 pag."
-              checked={yearList.includes("2023g")}
-              onChange={() => toggleYearInList("2023g")}
-            />
-            <Form.Check
-              inline
-              label="2023 pak."
-              checked={yearList.includes("2023k")}
-              onChange={() => toggleYearInList("2023k")}
-            />
+        <div style={{ marginTop: "20px", display: "flex" }}>
+          <Form style={{ flexGrow: 3 }}>
+            {allYearList.map((year) => (
+              <Form.Check
+                key={year}
+                inline
+                label={`${year.slice(0, 4)} pa${year.slice(4, 5)}.`}
+                checked={yearList.includes(year)}
+                onChange={() => toggleYearInList(year)}
+              />
+            ))}
           </Form>
+          <ShuffleBar
+            isShuffleOn={isShuffleOn}
+            setShuffleOn={setShuffleOn}
+            reshuffle={reshuffle}
+            undoShuffle={undoShuffle}
+            style={{ flexGrow: 1 }}
+          />
         </div>
+
         <hr />
       </div>
-
       <Accordion>
         {topics.map((topic) => (
-          <TopicSection key={topic.topic} topic={topic} yearList={yearList} />
+          <TopicSection
+            key={topic.topic}
+            topic={topic}
+            yearList={yearList}
+            isShuffleOn={isShuffleOn}
+          />
         ))}
       </Accordion>
     </div>