+++ src/misc.ts
@@ -40,3 +40,23 @@ export function parseMathProblemId(str: string) {
     isBlevel: isBLevel,
   };
 }
+
+export function shuffle(array: any[], isShuffleOn: boolean) {
+  if (isShuffleOn) {
+    let currentIndex = array.length,
+      randomIndex;
+
+    while (currentIndex > 0) {
+      randomIndex = Math.floor(Math.random() * currentIndex);
+      currentIndex--;
+
+      [array[currentIndex], array[randomIndex]] = [
+        array[randomIndex],
+        array[currentIndex],
+      ];
+    }
+    return array;
+  } else {
+    return array;
+  }
+}